Fix UniqueId parsing
Prior to this commit a type or value string containing a meta-character
used by the string representation syntax broke the parsing of an unique
id string.
Now, the parser uses 3 characters (defaults to `]/[`) to split the
segments of the unique id string representation. This improvement does
handle a lot more cases than the old implementation, but it can still
yield format exceptions when the type or value string contains exactly
the 3 characters in order.

Fixes #810